+++ R/bayesian_make.R
@@ -314,12 +314,14 @@ bayesian_make <- function(modes = c("classification", "regression")) {
               rlang::warn('Use `type = "raw"` for multivariate predictions!')
               return(results)
             }
-
+            threshold <- getOption("class_pred.threshold", 0.5)
             if (length(object$lvl) == 2) {
               if (is.array(results)) {
                 results <- as.vector(results)
               }
-              threshold <- getOption("class_pred.threshold", 0.5)
+              if (length(threshold) != 1) {
+                rlang::abort("Probability threshold should be a single value.")
+              }
               if (is.numeric(threshold)) {
                 if (!dplyr::between(threshold, 0, 1)) {
                   rlang::abort("Probability threshold is out of 0-1 range.")
@@ -336,6 +338,9 @@ bayesian_make <- function(modes = c("classification", "regression")) {
               length(object$lvl) > 2 &
                 length(object$lvl) == ncol(results)
             ) {
+              if (length(threshold) == ncol(results)) {
+                results <- sweep(results, 2, threshold, FUN = "/")
+              }
               results <- object$lvl[apply(results, 1, which.max)]
             } else {
               rlang::abort("Unexpected model predictions!")