+++ NEWS.md
@@ -3,6 +3,7 @@
 - Using base R pipe and depends on R >= 4.1.0
 - Using `parsnip::update_spec()` if exists
 - Supported the use of case weights
+- Supported threshold (inverse probability weights) for multi-class predictions
 - Fixed inconsistent probability (`type = "prob"`) predictions for two-class models
 - Updated package dependencies
 - Used `init` instead of the deprecated `inits` (`brms >= 2.16.9`)